requirements.txt
flask==3.0.3
flask-cors==4.0.0
PyPDF2==3.0.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
google-generativeai==0.7.2
groq==0.9.0
numpy==1.26.4